Moved from_unixtime function to gdv_stubs_func to use std libs
