+++ cpp/src/gandiva/gdv_function_stubs_test.cc
@@ -623,4 +623,94 @@ TEST(TestGdvFnStubs, TestInitCap) {
                   "unexpected byte \\e0 encountered while decoding utf8 string"));
   ctx.Reset();
 }
+
+TEST(TestGdvFnStubs, TestFromUnixtimeWithoutPattern) {
+  ExecutionContext context;
+  int64_t context_ptr = reinterpret_cast<int64_t>(&context);
+  gdv_int32 out_len;
+  // "2010-06-01 06:58:22" in seconds from epoch are:
+  gdv_timestamp ts = 1275375502;
+  const char* out = gdv_fn_from_unixtime_int64(context_ptr, ts, &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-06-01 06:58:22");
+
+  out = gdv_fn_from_unixtime_int64(context_ptr, 0L, &out_len);
+  EXPECT_EQ(std::string(out, out_len), "1970-01-01 00:00:00");
+
+  out = gdv_fn_from_unixtime_int64(context_ptr, -1L, &out_len);
+  EXPECT_EQ(std::string(out, out_len), "1969-12-31 23:59:59");
+}
+
+TEST(TestGdvFnStubs, TestFromUnixtimeWithPattern) {
+  ExecutionContext context;
+  int64_t context_ptr = reinterpret_cast<int64_t>(&context);
+  gdv_int32 out_len = 0;
+
+  // "2010-06-01 06:58:22" in seconds from epoch are:
+  gdv_timestamp ts_from_string =  1275375502;
+  const char* pattern1 = "yyyy-MM-dd hh:mm:ss";
+  int32_t pattern1_len = strlen(pattern1);
+  const char* out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern1,
+                                             pattern1_len, &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-06-01 06:58:22");
+
+  const char* pattern2 = "yyyy-MM-dd hh:mm";
+  int32_t pattern2_len = strlen(pattern2);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern2, pattern2_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-06-01 06:58");
+
+  const char* pattern3 = "yyyy-MM-dd hh";
+  int32_t pattern3_len = strlen(pattern3);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern3, pattern3_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-06-01 06");
+
+  const char* pattern4 = "YYYY-MM-dd";
+  int32_t pattern4_len = strlen(pattern4);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern4, pattern4_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-06-01");
+
+  const char* pattern5 = "YYYY-dd-MM";
+  int32_t pattern5_len = strlen(pattern5);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern5, pattern5_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "2010-01-06");
+
+  const char* pattern6 = "dd-M-yy hh:mm";
+  int32_t pattern6_len = strlen(pattern6);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern6, pattern6_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "01-Jun-10 06:58");
+
+  const char* pattern7 = "Mm-dd-YY hh:mm:ss";
+  int32_t pattern7_len = strlen(pattern7);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern7, pattern7_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "June-01-10 06:58:22");
+
+  const char* pattern8 = "hh:mm:ss dd-M-yyyy";
+  int32_t pattern8_len = strlen(pattern8);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, pattern8, pattern8_len,
+                                 &out_len);
+  EXPECT_EQ(std::string(out, out_len), "06:58:22 01-Jun-2010");
+
+  const char* invalid_pattern1 = "yyyy-MM-dd hh:mm:ss.ss";
+  int32_t invalid_pattern1_len = strlen(invalid_pattern1);
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, invalid_pattern1,
+                                 invalid_pattern1_len, &out_len);
+  std::string out_str = std::string(out, out_len);
+  EXPECT_EQ(out_str, "");
+  EXPECT_EQ(context.get_error(), "Pattern has invalid delimiter '.'");
+  context.Reset();
+
+  const char* invalid_pattern3 = "";
+  int32_t invalid_pattern3_len = -1;
+  out = gdv_fn_from_unixtime_int64_utf8(context_ptr, ts_from_string, invalid_pattern3,
+                                 invalid_pattern3_len, &out_len);
+  out_str = std::string(out, out_len);
+  EXPECT_EQ(out_str, "");
+  EXPECT_EQ(context.get_error(), "Invalid pattern, it must have at least 1 char");
+  context.Reset();
+}
 }  // namespace gandiva
